requirements.txt
autopep8==1.5.2
boto3==1.12.46
botocore==1.15.46
docutils==0.15.2
jmespath==0.9.5
pycodestyle==2.5.0
python-dateutil==2.8.1
python-dotenv==0.13.0
s3transfer==0.3.3
six==1.14.0
urllib3==1.25.9
